Launcher3: Close hotseat folder when icon dragging stops
This prevents a force close when removing an app from a folder with just one app inside, by ensuring those are immediately replaced by the app itself. The cause of this is that folders on the hotseat are opened outside of the opening drag bounds. The closing bounds are not passed when dragging the icon away again, keeping the folder open with potentially only one app inside.